make memory pool creation self-contained in memory.rs
